backend/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
archspec @ file:///croot/archspec_1709217642129/work
asgiref @ file:///C:/b/abs_ae1n0p_8_w/croot/asgiref_1724072476516/work
attrs==23.2.0
blinker==1.9.0
boltons @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/boltons_1699480450092/work
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
conda @ file:///C:/b/abs_8erkunzmxm/croot/conda_1732198201855/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_bdz9vbvbh2/croot/conda-package-streaming_1731366449946/work
cssselect2==0.7.0
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
Django @ file:///C:/b/abs_d8jy6px5xv/croot/django_1731616483677/work
django-cors-headers==4.6.0
djangorestframework==3.15.2
easyocr==1.7.1
filelock==3.15.4
Flask==3.1.0
fonttools==4.53.1
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.34.2
importlib-metadata==6.11.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
libmambapy @ file:///C:/b/abs_7fjjcok9fe/croot/mamba-split_1732896315069/work/libmambapy
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
mdurl==0.1.2
menuinst @ file:///C:/b/abs_cabvzz4p7a/croot/menuinst_1731364936347/work
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
openai==1.35.14
opencv-python-headless==4.10.0.84
packaging==23.2
pandas==2.2.2
pdfkit==1.0.0
pillow==10.4.0
platformdirs @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/platformdirs_1701797392447/work
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
protobuf==4.25.3
pyarrow==17.0.0
pyclipper==1.3.0.post5
pycosat @ file:///C:/b/abs_5csdern___/croot/pycosat_1714513102923/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
pyphen==0.15.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
reportlab==4.2.2
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rich==13.7.1
rpds-py==0.19.0
ruamel.yaml @ file:///C:/b/abs_0cunwx_ww6/croot/ruamel.yaml_1727980181547/work
ruamel.yaml.clib @ file:///C:/b/abs_5fk8zi6n09/croot/ruamel.yaml.clib_1727769837359/work
scikit-image==0.24.0
scipy==1.14.0
setuptools==75.1.0
shapely==2.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/sqlparse_1699544474746/work
streamlit==1.27.1
sympy==1.13.1
tbb==2021.13.0
tenacity==8.5.0
tifffile==2024.7.2
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
truststore @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/truststore_1701881385424/work
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
validators==0.33.0
watchdog==4.0.1
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.44.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
zipp==3.19.2
zopfli==0.2.3
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
